New script: WME Form Filler
WME Form Filler used to create links to fill forms with data from WME.
This is a basic version only used to support the USA Emergencies team
and is built solely for that purpose right now. Hopefully, it's not
so horribly done that it's very hard to expand to allow users to
manually add their own forms within WME in the future.